% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_integ.R
\name{allele_integ}
\alias{allele_integ}
\title{feature-level allele imbalance integration}
\usage{
allele_integ(hbomics, phased, feature, ...)
}
\arguments{
\item{hbomics}{The hbomics object from HBomics function,}

\item{phased}{phased SNV file with two columns: ID contain SNV id in the HBomics object;
GT column contain phase information such as "1|0" or "0|1"}

\item{feature}{indicate the feature column in the HBomics object hb_df data frame;}
}
\value{
An \code{HBomicsAlleleResult} object with feature-level summaries in
\code{gene_allele_imb} and the integrated site-level data in \code{HBomics}
}
\description{
Function for allele imbalance ratio integration to feature-level (gene/peak)
with user provided phase information
for feature with multiple SNVs, allele imbalance ratios are weighted by the allele depth from the enrichment data,
i.e., the most expressed allele weights more on the integrative allele imbalance ratios
}
\details{
HBomics algorithm: calling clonal allele imbalance events from multiomics data with a hierarchical Bayesian framework
Author: Zhen Y
}
