% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_kl_divergence}
\alias{calc_kl_divergence}
\title{Compute KL Divergence between two MCMC posterior samples (sample-wise comparison)}
\usage{
calc_kl_divergence(c_samples, t_samples, n_grid = 512)
}
\arguments{
\item{n_grid}{Number of points to evaluate the density over (default 512)}

\item{p_samples}{Numeric vector of MCMC samples from distribution P}

\item{q_samples}{Numeric vector of MCMC samples from distribution Q}
}
\value{
Numeric KL divergence value
}
\description{
Compute KL Divergence between two MCMC posterior samples (sample-wise comparison)
}
