% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CI_quant}
\alias{CI_quant}
\title{Get Confidence/Credible Interval boundaries of the AI posterior distribution}
\usage{
CI_quant(x, bound = 0.95)
}
\arguments{
\item{x}{A numeric vector of MCMC samples}

\item{bound}{The interval boundary (default is 0.95)}
}
\value{
A vector with the lower and upper quantiles
}
\description{
Get Confidence/Credible Interval boundaries of the AI posterior distribution
}
