% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBomics_KL.R
\name{compare_hbomics_kl}
\alias{compare_hbomics_kl}
\title{KL divergence of sample-wise comparison}
\usage{
compare_hbomics_kl(obj_C, obj_T, symmetric = TRUE)
}
\arguments{
\item{obj_C}{First fitted HBomics result object (e.g., control or primary).}

\item{obj_T}{Second fitted HBomics result object (e.g., treatment).}

\item{symmetric}{Logical; if TRUE, calculates Jeffrey's divergence (default TRUE)}
}
\value{
A data frame with class \code{HBomicsKLResult}, sorted by descending KL divergence
}
\description{
This function uses posterior samples from HBomics object for site-by-site comparison of different samples
It also provides feature-by-feature comparison with allele depth weighted KL scores after
running allele_integ function
}
\details{
HBomics algorithm: calling clonal allele imbalance events from multiomics data with a hierarchical Bayesian framework
Author: Zhen Y
}
