% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_list_d}
\alias{data_list_d}
\title{Stan Data List of COLO320-DM}
\format{
A list with elements required by the Stan model (T, C, S, A, N, etc.)
}
\usage{
data_list_d
}
\description{
Stan Data List of COLO320-DM
}
\keyword{datasets}
