% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBomics-package.R
\docType{package}
\name{HBomics-package}
\alias{HBomics-package}
\alias{HBomics}
\title{The 'HBomics' package.}
\description{
A hierarchical Bayesian framework for clonal allele imbalance inference from tumor genome using WGS combined with other multiomics data such as RNA-seq and ChIP-seq
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Zhen You \email{youzhen93@gmail.com} (\href{https://orcid.org/0000-0002-0578-4823}{ORCID})

}
