% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBomics.R
\name{HBomics}
\alias{HBomics}
\title{HBomics Class}
\description{
R6 class for fitting the HBomics model and retrieving validated result objects.
}
\details{
HBomics algorithm: calling clonal allele imbalance events from multiomics data with a hierarchical Bayesian framework
Author: Zhen Y
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HBomics-new}{\code{HBomics$new()}}
\item \href{#method-HBomics-fit_model}{\code{HBomics$fit_model()}}
\item \href{#method-HBomics-convergence}{\code{HBomics$convergence()}}
\item \href{#method-HBomics-result}{\code{HBomics$result()}}
\item \href{#method-HBomics-get_fit}{\code{HBomics$get_fit()}}
\item \href{#method-HBomics-get_diagnostics}{\code{HBomics$get_diagnostics()}}
\item \href{#method-HBomics-get_hb_df}{\code{HBomics$get_hb_df()}}
\item \href{#method-HBomics-get_posterior_samples}{\code{HBomics$get_posterior_samples()}}
\item \href{#method-HBomics-clone}{\code{HBomics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-new"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$new(data_list, mutation_id, meta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-fit_model"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-fit_model}{}}}
\subsection{Method \code{fit_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$fit_model(
  chains = 4,
  warmup = 1000,
  iter = 2000,
  cores = 1,
  seed = 123,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-convergence"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-convergence}{}}}
\subsection{Method \code{convergence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$convergence()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-result"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-result}{}}}
\subsection{Method \code{result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$result(KL = FALSE, bound = 0.95)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-get_fit"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-get_fit}{}}}
\subsection{Method \code{get_fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$get_fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-get_diagnostics"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-get_diagnostics}{}}}
\subsection{Method \code{get_diagnostics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$get_diagnostics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-get_hb_df"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-get_hb_df}{}}}
\subsection{Method \code{get_hb_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$get_hb_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-get_posterior_samples"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-get_posterior_samples}{}}}
\subsection{Method \code{get_posterior_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$get_posterior_samples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HBomics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HBomics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HBomics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
