% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_d}
\alias{meta_d}
\title{HBomics dummy data set}
\format{
A metasheet data frame with 1412 rows and 13 variables from COLO320-DM:
}
\source{
For tutorial purposes.
}
\usage{
meta_d
}
\description{
example clonal SNVs data from COLO320-DM/HSR treated with STP,
used to demonstrate the HBomics pipeline.
}
\keyword{datasets}
