% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode2}
\alias{Mode2}
\title{HBomics algorithm: calling clonal allele imbalance events from multiomics data with a hierarchical Bayesian framework
Author: Zhen Y
Utils function
Calculate the allele imbalance ratio (the summit) from the AI posterior distribution}
\usage{
Mode2(x)
}
\arguments{
\item{x}{A numeric vector of MCMC samples}
}
\value{
The inferred allele imbalance ratio
}
\description{
HBomics algorithm: calling clonal allele imbalance events from multiomics data with a hierarchical Bayesian framework
Author: Zhen Y
Utils function
Calculate the allele imbalance ratio (the summit) from the AI posterior distribution
}
