% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phased}
\alias{phased}
\title{Phased SNV Data from eagle2}
\format{
A data frame with 1412 rows and 2 variables:
}
\usage{
phased
}
\description{
Phased SNV Data from eagle2
}
\keyword{datasets}
